adds references to other materials for explaining git concepts.
It was a little hard to justify re-writing everything about how to use git when there's already so much good information out there.
I opted to instead just refer to materials that I've used in my learning, and contextualize them a little to help folks end up in the right place.
With this pull request, I think the workshop is ready for some feedback.